+++ tests/sdk/test_update.py
@@ -19,24 +19,24 @@ from .. import datastore
 from ..utils import mock_requests
 
 
-def test_update_compute_plan(client, mocker):
+def test_add_compute_plan_tuples(client, mocker):
     item = datastore.COMPUTE_PLAN
     m = mock_requests(mocker, "post", response=item)
     m_get = mock_requests(mocker, "get", response=datastore.COMPUTE_PLAN)
 
-    response = client.update_compute_plan("foo", {})
+    response = client.add_compute_plan_tuples("foo", {})
 
     assert response == models.ComputePlan(**item)
     m.assert_called
     m_get.assert_called
 
 
-def test_update_compute_plan_with_schema(client, mocker):
+def test_add_compute_plan_tuples_with_schema(client, mocker):
     item = datastore.COMPUTE_PLAN
     m = mock_requests(mocker, "post", response=item)
     m_get = mock_requests(mocker, "get", response=datastore.COMPUTE_PLAN)
 
-    response = client.update_compute_plan("foo", schemas.UpdateComputePlanSpec(key="foo"))
+    response = client.add_compute_plan_tuples("foo", schemas.UpdateComputePlanSpec(key="foo"))
 
     assert response == models.ComputePlan(**item)
     m.assert_called